requirements.txt
# Core GUI/runtime dependencies.
PyQt6>=6.11.0
openai
requests
rapidfuzz
Pillow>=12.2.0
psd-tools
Unidecode

# Optional feature modules used when available. The app also bootstraps these
# at runtime, but keeping them installable here makes release builds and fresh
# developer environments deterministic.
pillow-heif
exifread
mutagen
pypdf
python-docx
openpyxl
python-pptx
reverse_geocoder
send2trash
# rarfile / py7zr extract untrusted archives.  Pair with
# fileorganizer.safe_archive.safe_extract_path to block path traversal in
# archive entry names.  Two open GitHub Advisory DB entries each (N-13).
rarfile
py7zr
# fonttools — pinned to fix CVE-2025-66034 (path traversal in varLib.main,
# fixed in 4.61.0).  N-9 metadata extractors will use the name table; the
# pin lands here ahead of that work to cover any transitive use today.
fonttools>=4.62.1

# Optional local embeddings classifier (N-10).  Install at least one to enable
# the pre-AI cosine match against category anchors; the classifier degrades to
# an AI-only flow when none is present.
# fastembed is the preferred backend (ONNX runtime, ~80 MB on disk).
# fastembed
# model2vec
# sentence-transformers
